Remove unneeded files.
