+++ src/nextapp/components/service-accounts-list/index.ts
@@ -1,0 +1,1 @@
+export { default } from './service-accounts-list';
